Move `deltaExtensions` related methods up to `AbstractExtensionService`
